network-script: upgrade install-network-script.sh to take ROLE input

+++ test-dut-network.sh
@@ -3,15 +3,18 @@
 source ./network-script/setup-global.sh
 source ./network-script/helper-lib.sh
 
+alive=$(alive_test)
+if [ x"$alive" != x"PASS" ]; then
+        echo "ABORT: $alive"
+        exit 0
+fi
+
 # Delete previous authorized public keys
 rm ${TC_HOME}/.ssh/authorized_keys
 
 # Update all network scripts in both DUT & LP
-./install-network-script.sh
-
-# Setup DUT & LP
-run_dut "cd ${NETSCRIPT_INSTALL}; ./setup-dut.sh"
-run_lp "cd ${NETSCRIPT_INSTALL}; ./setup-lp.sh"
+./install-network-script.sh DUT
+./install-network-script.sh LP
 
 TEST_TIME=$(print_time)
 echo -e "Test runs at ${TEST_TIME}"
